// images/frontend/app_r18/src/contexts/platformAssistantContext/interfaces.tsx
export interface PlatformAssistantDispatch {
	(arg0: { type: string; payload?: any; error?: any }): void;
}
// ...
export interface IUserRole {
	userRole: string;
	numOrganizationManaged: number;
	numGroupsManaged: number;
	numAssetTypesManaged: number;
	numAssetsManaged: number;
	numSensorsManaged: number;
	numDigitalTwinsManaged: number;
	numMLModelsManaged: number;
}

export interface IPlatformAssistantOptionToShow {
	platformAssistantOptionToShow: string;
}

export interface ISelectOrgUsersTable {
	selectOrgUsers: ISelectOrgUser[];
}

export interface IReloadSelectOrgUsersTable {
	reloadSelectOrgUsersTable: boolean;
}

export interface IOrganizationsTable {
	organizations: IOrganization[];
}

export interface INodeRedInstancesTable {
	nodeRedInstances: INodeRedInstance[];
}

export interface IReloadNodeRedInstancesTable {
	reloadNodeRedInstancesTable: boolean;
}

export interface IGlobalUsersTable {
	globalUsers: IGlobalUser[];
}

export interface IReloadGlobalUsersTable {
	reloadGlobalUsersTable: boolean;
}

export interface IBuildingsTable {
	buildings: IBuilding[];
}

export interface IReloadBuildingsTable {
	reloadBuildingsTable: boolean;
}

export interface IFloorsTable {
	floors: IFloor[];
}

export interface IReloadFloorsTable {
	reloadFloorsTable: boolean;
}

export interface IRefreshTokensTable {
	refreshTokens: IRefreshToken[];
}

export interface IOrgsManagedTable {
	orgsManaged: IOrgManaged[];
}

export interface IReloadOrgsManagedTable {
	reloadOrgsManagedTable: boolean;
}

export interface IOrgsOfGroupsManagedTable {
	orgsOfGroupsManaged: IOrgOfGroupsManaged[];
}

export interface IReloadOrgsOfGroupsManagedTable {
	reloadOrgsOfGroupsManagedTable: boolean;
}

export interface IOrgUsersTable {
	orgUsers: IOrgUser[];
}

export interface IReloadOrgUsersTable {
	reloadOrgUsersTable: boolean;
}

export interface IGroupsTable {
	groups: IGroup[];
}

export interface IReloadGroupsTable {
	reloadGroupsTable: boolean;
}

export interface IGroupsManagedTable {
	groupsManaged: IGroupManaged[];
}

export interface IReloadGroupsManagedTable {
	reloadGroupsManagedTable: boolean;
}

export interface IGroupMembersTable {
	groupMembers: IGroupMember[];
}

export interface IReloadGroupMembersTable {
	reloadGroupMembersTable: boolean;
}

export interface IAssetTypesTable {
	assetTypes: IAssetType[];
}

export interface IReloadAssetTypesTable {
	reloadAssetTypesTable: boolean;
}

export interface IAssetsTable {
	assets: IAsset[];
}

export interface IReloadAssetsTable {
	reloadAssetsTable: boolean;
}

export interface ISensorsTable {
	sensors: ISensor[];
}

export interface IReloadSensorsTable {
	reloadSensorsTable: boolean;
}

export interface IUserProfileTable {
	userProfile: IUserProfile;
}

export interface IOrgsMembershipTable {
	orgsMembership: IMembershipInOrgs[];
}

export interface IRealoadOrgsMembershipTable {
	reloadOrgsMembershipTable: boolean; 
}

export interface IGroupsMembershipTable {
	groupsMembership: IMembershipInGroups[];
}

export interface IReloadGroupsMembershipTable {
	reloadGroupsMembershipTable: boolean;
}

export interface ITopicsTable {
	topics: ITopic[];
}

export interface IReloadTopicsTable {
	reloadTopicsTable: boolean;
}


export interface IDashboardsTable {
	dashboards: IDashboard[];
}

export interface IReloadDashboardsTable {
	reloadDashboardsTable: boolean;
}

export interface IDigitalTwinsTable {
	digitalTwins: IDigitalTwin[];
}

export interface IReloadDigitalTwinsTable {
	reloadDigitalTwinsTable: boolean;
}

export interface IMlModelsTable {
	mlModels: IMlModel[];
}

export interface IReloadMlModelsTable {
	reloadMlModelsTable: boolean;
}

export interface IWindowObjectReferences {
	windowObjectReferences: Record<string, Window | null>;
}